* After #93 ,`insert = union (one x)` gives O(n) inserts, still not great; but better than `O(nlogn)`
* `delete`: Using `toSortedTriples`, which is an `O(n)` way to get a sorted list of triples, we can simply filter out the offending element and rebuild in `O(n)` total time

Addresses https://github.com/kowainik/typerep-map/issues/57

This should be rebased on top of #93  once it's merged,

In the meantime; [these are the relevant commits from just this PR](https://github.com/kowainik/typerep-map/pull/94/files/cb082cd3a8968c6894d44a0510b740dd55290dc3..1fc35afde2eda4f6c957af34e06f20b4395a845f)