@ChrisPenner 2020-10-11
This PR has been rebased and is ready for review 👍