@ChrisPenner 2020-10-11
I'll run some benchmarking on this too and report back.