@ChrisPenner 2020-10-11
Here're the relevant comparisons between current master and this branch:

# Insert
**master**
```
benchmarking insert/10 elements to empty/CacheMap
time                 50.55 μs   (47.29 μs .. 53.36 μs)
                     0.974 R²   (0.966 R² .. 0.985 R²)
mean                 49.18 μs   (47.44 μs .. 52.17 μs)
std dev              6.973 μs   (6.348 μs .. 8.210 μs)
variance introduced by outliers: 91% (severely inflated)

benchmarking insert/1 element to big map/CacheMap
time                 334.6 ms   (311.4 ms .. 353.3 ms)
                     0.999 R²   (NaN R² .. 1.000 R²)
mean                 341.2 ms   (337.2 ms .. 345.6 ms)
std dev              5.298 ms   (1.587 ms .. 7.033 ms)
variance introduced by outliers: 19% (moderately inflated)
```

**new**
```
benchmarking insert/10 elements to empty/CacheMap
time                 46.18 μs   (43.57 μs .. 49.69 μs)
                     0.973 R²   (0.963 R² .. 0.984 R²)
mean                 47.63 μs   (45.77 μs .. 50.72 μs)
std dev              8.029 μs   (6.576 μs .. 10.69 μs)
variance introduced by outliers: 94% (severely inflated)

benchmarking insert/1 element to big map/CacheMap
time                 9.916 ms   (9.607 ms .. 10.22 ms)
                     0.995 R²   (0.992 R² .. 0.997 R²)
mean                 9.788 ms   (9.682 ms .. 9.989 ms)
std dev              395.2 μs   (335.5 μs .. 545.5 μs)
variance introduced by outliers: 16% (moderately inflated)

```